function contracts are live, calls remain TODO

+++ rust/src/control_flow/block_id.rs
@@ -5,10 +5,10 @@ pub struct BlockID {
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct FunctionID {
-    pub id: usize, // yes, we could just use usize directly, this is more clear and costs nothing
+    pub id: i64, // yes, we could just use usize directly, this is more clear and costs nothing
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct ClassID {
-    pub id: usize, // yes, we could just use usize directly, this is more clear and costs nothing
+    pub id: i64, // yes, we could just use usize directly, this is more clear and costs nothing
 }
